RBAC with CASL

+++ app/categories/controller.js
@@ -1,7 +1,17 @@
 const Category = require("./model");
+const { policyFor } = require("../policy");
 
 async function destroy(req, res, next) {
   try {
+    let policy = policyFor(req.user);
+
+    if (!policy.can("delete", "Category")) {
+      return res.json({
+        error: 1,
+        message: "Anda tidak memiliki akses untuk menghapus kategori",
+      });
+    }
+
     let category = await Category.findOneAndDelete({ _id: req.params.id });
 
     return res.json(category);
@@ -12,6 +22,15 @@ async function destroy(req, res, next) {
 
 async function update(req, res, next) {
   try {
+    let policy = policyFor(req.user);
+
+    if (!policy.can("update", "Category")) {
+      return res.json({
+        error: 1,
+        message: "Anda tidak memiliki akses untuk mengupdate kategori",
+      });
+    }
+
     let payload = req.body;
     let category = await Category.findByIdAndUpdate(
       { _id: req.params.id },
@@ -39,6 +58,15 @@ async function update(req, res, next) {
 
 async function store(req, res, next) {
   try {
+    let policy = policyFor(req.user);
+
+    if (!policy.can("create", "Category")) {
+      return res.json({
+        error: 1,
+        message: "Anda tidak memiliki akses untuk membuat kategori",
+      });
+    }
+
     let payload = req.body;
     let category = new Category(payload);
     await category.save();
